requirements.txt
python-jenkins
requests
boto3